% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comb2BLRM_EXNEX.R
\name{Combo2BLRM.EXNEX}
\alias{Combo2BLRM.EXNEX}
\title{Combo2BLRM.EXNEX Function}
\usage{
Combo2BLRM.EXNEX(
  agent1 = NULL,
  agent2 = NULL,
  DoseRef1 = NULL,
  Prior.NEX.Agent1 = NULL,
  Prior.EX.mu1.Agent1 = NULL,
  Prior.EX.mu2.Agent1 = NULL,
  Prior.EX.tau1.Agent1 = NULL,
  Prior.EX.tau2.Agent1 = NULL,
  Prior.rho.Agent1 = c(-1, 1),
  DoseRef2 = NULL,
  Prior.NEX.Agent2 = NULL,
  Prior.EX.mu1.Agent2 = NULL,
  Prior.EX.mu2.Agent2 = NULL,
  Prior.EX.tau1.Agent2 = NULL,
  Prior.EX.tau2.Agent2 = NULL,
  Prior.rho.Agent2 = c(-1, 1),
  Prior.EX.eta.mu = NULL,
  Prior.EX.eta.tau = NULL,
  Prior.NEX.eta = NULL,
  Pint = c(0.16, 0.33),
  MCMC = c(2500, 7500, 4, 1),
  Rnd.seed = 1452671,
  RSeed.WinBUGS = 1367,
  DIC = TRUE,
  progress.bar = "text",
  pj.EX = NULL,
  Doses1 = NULL,
  Doses2 = NULL,
  trialData = NULL
)
}
\arguments{
\item{agent1}{is a character string with the name/abbreviation for agent 1.}

\item{agent2}{is a character string with the name/abbreviation for agent 2.}

\item{DoseRef1}{refernce dose for agent 1 used in the BLRM.}

\item{Prior.NEX.Agent1}{is the nonexchangeability prior for agent 1 used in the combination BLRM, \eqn{log(\alpha)} and \eqn{log(\beta)}. This is a vector of length 5 for the NEX bivariate normal prior for, (\eqn{m_{log(\alpha)}}, \eqn{m_{log(\beta)}}, \eqn{s_{log(\alpha)}}, \eqn{s_{log(\beta)}}, \eqn{Corr(log(\alpha),log(\beta))}).}

\item{Prior.EX.mu1.Agent1}{is the EX prior specification for Agent 1 for the \eqn{log(\alpha)} mean parameter. It is a vector of length 2, (\eqn{m_{\mu1}, s_{\mu1}}).}

\item{Prior.EX.mu2.Agent1}{is the EX prior specification for Agent 1 for the \eqn{log(\beta)} mean parameter. It is a vector of length 2, (\eqn{m_{\mu2}, s_{\mu2}}).}

\item{Prior.EX.tau1.Agent1}{is the EX prior specification for Agent 1 for the \eqn{log(\alpha)} standard deviation parameter. It is a vector of length 2, (\eqn{m_{\tau1}, s_{\tau1}}).}

\item{Prior.EX.tau2.Agent1}{is the EX prior specification for Agent 1 for the \eqn{log(\beta)} standard deviation parameter. It is a vector of length 2, (\eqn{m_{\tau2}, s_{\tau2}}).}

\item{Prior.rho.Agent1}{is the uniform prior specification for Agent 1 for the correlation between \eqn{\tau1} and \eqn{\tau2}. It is a vector of length 2, and the default is set to (-1,1).}

\item{DoseRef2}{refernce dose for agent 2 used in the BLRM.}

\item{Prior.NEX.Agent2}{is the nonexchangeability prior for agent 2 used in the combination BLRM, \eqn{log(\alpha)} and \eqn{log(\beta)}. This is a vector of length 5 for the NEX bivariate normal prior for, (\eqn{m_{log(\alpha)}}, \eqn{m_{log(\beta)}}, \eqn{s_{log(\alpha)}}, \eqn{s_{log(\beta)}}, \eqn{Corr(log(\alpha),log(\beta))}).}

\item{Prior.EX.mu1.Agent2}{is the EX prior specification for Agent 2 for the \eqn{log(\alpha)} mean parameter. It is a vector of length 2, (\eqn{m_{\mu1}, s_{\mu1}}).}

\item{Prior.EX.mu2.Agent2}{is the EX prior specification for Agent 2 for the \eqn{log(\beta)} mean parameter. It is a vector of length 2, (\eqn{m_{\mu2}, s_{\mu2}}).}

\item{Prior.EX.tau1.Agent2}{is the EX prior specification for Agent 2 for the \eqn{log(\alpha)} standard deviation parameter. It is a vector of length 2, (\eqn{m_{\tau1}, s_{\tau1}}).}

\item{Prior.EX.tau2.Agent2}{is the EX prior specification for Agent 2 for the \eqn{log(\beta)} standard deviation parameter. It is a vector of length 2, (\eqn{m_{\tau2}, s_{\tau2}}).}

\item{Prior.rho.Agent2}{is the uniform prior specification for Agent 2 for the correlation between \eqn{\tau1} and \eqn{\tau2}. It is a vector of length 2, and the default is set to (-1,1).}

\item{Prior.EX.eta.mu}{is the mean EX prior specification for the interaction parameter, i.e., log odds multiplier \eqn{\eta}. It is a vector of length 2, (\eqn{m_{\eta}, s_{\eta}}).}

\item{Prior.EX.eta.tau}{is the standard deviation EX prior specification for the interaction parameter, i.e., log odds multiplier \eqn{\eta}. It is a vector of length 2, (\eqn{m_{\tau}, s_{\tau}}).}

\item{Prior.NEX.eta}{is the nonexchangeability prior specification for the interaction parameter, i.e., log odds multiplier \eqn{\eta}. It is a vector of length 2, (\eqn{m_{\eta}, s_{\eta}}).}

\item{Pint}{targeted interval probability cutoff, which will be used to define three intervals for the true DLT probability. The default is
(0.16, 0.33) defining three categories: underdosing ({[}0--0.16{]}), target ((0.16--0.33)), and 
overdosing ({[}0.33--1.00{]}).}

\item{MCMC}{four component vector of MCMC parameters n.burnin, n.iterations, n.chains, n.thin. See \link{R2WinBUGS} for details.}

\item{Rnd.seed}{Random seed for the R random stream. This is important for reproducibility.}

\item{RSeed.WinBUGS}{Random seed for the WinBUGS random stream. This is important for reproducibility.}

\item{DIC}{logical: if TRUE, compute deviance, pD, and DIC. See \link{R2WinBUGS} for details.}

\item{progress.bar}{type of progress bar. Possible values are "text", "gui", and "none". Type "text"
is displayed on the R console. See \link{rjags} for details.}

\item{pj.EX}{is the proportion of exchangeability, resulting in a 1- pj.EX nonexchangeability proportion. For example, pj.EX = 1 is full exchangeability, while pj.EX=0 is a stratified analysis.}

\item{Doses1}{a vector of doses for agent 1 for which inferential results will be provided, i.e., the provisional dose levels.}

\item{Doses2}{a vector of doses for agent 2 for which inferential results will be provided, i.e., the provisional dose levels.}

\item{trialData}{dose-DLT data: a list with length equal to the number of trials. Data for each trial is entered as a list object with 5 elements; DoseAdm1,DoseAdm2, Npat, Ntox, trialName. DoseAdm1 is a vector of the doses administered for agent 1. DoseAdm2 is a vector of the doses administered for agent 2. Npat are the number of participants treated at a given dose. Ntox are the number of DLTs observed for a given dose. trialName is the unique name used to identfy a specific trial.}
}
\value{
The function returns as a list the following main items:
\item{P}{Posterior summary of DLT probabilities for the different dose combinations}
\item{Pcat}{Posterior summary of DLT interval probabilities for the different dose combinations}
\item{PcatP}{P and Pcat combined}
\item{logAlphaBeta}{Posterior summary of (\eqn{\log(\alpha)}, \eqn{\log(\beta)}) for agent 1 and 2}
\item{PmixEXNEX}{Posterior mixture weigths}
\item{intervalPlot}{Interval plots for DLT rates. This is output as a list equal to the length of the number of trials/strata.}
\item{postProbPlot}{Estimates and 95\% interval plots for DLT rates. This is output as a list equal to the length of the number of trials/strata.}
}
\description{
This function analyzes dose limiting toxicity (DLT) data using a Bayesian logistic regression model (BLRM) combined with the Exchangeability-nonexchangeability (EXNEX) approach for 2 agent Phase I Oncology trials. The EXNEX approach allows information to be borrowed across concurrent trials.
}
\details{
Statistical model:
Details on the statistical model are provided below and additional details are provided in the references.

Prior Specification:
The Bayesian approach requires the specification of the prior distributions for the model parameters for each agent and their interaction, e.g., \eqn{log(\alpha)}, \eqn{log(\beta)}, and \eqn{log(\eta)}. These priors need to be specified for both EX and NEX.

EXNEX for binary data:
The EXNEX apporach allows the use of co-data where stratum specific (e.g. study specific) parameters can be either exchangeable with parameters from other strata or nonexchangeable. For binary data this is the log-odds parameter, \eqn{\theta_j = log(\frac{\pi_j}{1-\pi_j})}.

For each strata, j, we have weights \eqn{p_j} and \eqn{1-p_j}. Where the parameters are exchangeable (EX) with probability \eqn{p_j} and \eqn{\theta_j} follows a normal distribution with exchangeability parameters \eqn{\mu} and \eqn{\tau},

\eqn{\theta_j \vert \mu, \tau \sim N(\mu,\tau^2)}

and nonexchangeable (NEX) with probability \eqn{1-p_j} and strata-specific priors are used where,

\eqn{\theta_j \sim N(m_j,v_j)}.

Combination BLRM with 2 agents and EXNEX:

The dose-toxicity relationship will be modeled using a 5-parameter BLRM where for a combination dose of agent 1 \eqn{d_1^*} and agent 2 \eqn{d_2^*}, the number of participants with a dose limiting toxicity (DLT), \eqn{r_d}, in cohort of size \eqn{n_d} is binomial,

\eqn{r_d \vert \pi_d \sim Binomial(\pi_d, n_d)}

and for each stratum, j, and therapy the dose-toxicity model is logistic,

\eqn{logit(\pi_{1jd}) = log(\alpha_{1j}) + \beta_{1j} log(\frac{d_1}{d_1^*})}

\eqn{logit(\pi_{2jd}) = log(\alpha_{2j}) + \beta_{2j} log(\frac{d_2}{d_2^*})}

where \eqn{d_1} and \eqn{d_2} are the current combination doses under evaluation and \eqn{d_1^*} and \eqn{d_2^*} are the reference doses, which are used to scale the doses. The \eqn{\alpha, \beta > 0} are the parameters of the model such that \eqn{\alpha} is the therapy's odds of a DLT at \eqn{d^*}, and \eqn{\beta} is the increase in the log-odds of a DLT for a unit increase in the log-dose.

The dose-DLT relationship of the combination is defined as 

\eqn{Odds(\pi_{12}(d_1,d_2)) = \frac{\pi_{12}(d_1,d_2)}{1-\pi_{12}(d_1,d_2)}}
\eqn{ = exp(\eta \frac{d_1}{d_1^*}\frac{d_2}{d_2^*})[\frac{\pi_1(d_1)+\pi_2(d_2)-\pi_1(d_1)\pi_2(d_2)}{(1-\pi_1(d_1))(1-\pi_2(d_2))}]}

where \eqn{\eta} is the interaction paramter between agent 1 and agent 2. The Bayesian approach requires the specification of prior of prior distributions for the 5 model parameters \eqn{log(\alpha_{1})}, \eqn{log(\alpha_{2})}, \eqn{log(\beta_{1})}, \eqn{log(\beta_{2})}, and the interaction parameter \eqn{\eta}.

The model is then confined to one EX distribution and one NEX distribution, with *a priori* \eqn{p_j} and \eqn{1-p_j}. Under \eqn{p_j = 1} parameters are fully exchangeable, while \eqn{p_j = 0} would result in a fully stratified analysis with no exchangeability. Note currenty only one weight can be specified and it is applied acorss all strata, i.e., \eqn{j=1}.

With probability \eqn{p_j} for EX, the logistic parameter vector for agent 1

\eqn{\theta_{1j} = [log(\alpha_{1j}),log(\beta_{1j})]}

is exchangeable with the other strata parameters and follows a bivaraite normal distribution with mean vector \eqn{\mu_1} and covariance matrix \eqn{\Sigma_1};

\eqn{\theta_{1j} \sim BVN(\mu_{1},\Sigma_{1})}

where

\eqn{\mu_{1} = (\mu_{11}, \mu_{12})} 

and 

\eqn{\Sigma_1 = \left({\begin{array}{cc} \tau^2_{11} & \rho\tau_{11}\tau_{12} \\ \rho\tau_{11}\tau_{12} &  \tau^2_{12}\end{array}}\right)}.

the logistic parameter vector for agent 2,

#'\eqn{\theta_{2j} = [log(\alpha_2j),log(\beta_2j)]}

is exchangeable with the other strata parameters and follows a bivaraite normal distribution with mean vector \eqn{\mu_2} and covariance matrix \eqn{\Sigma_2};

\eqn{\theta_{2j} \sim BVN(\mu_{2},\Sigma_2)}

where

\eqn{\mu_2 = (\mu_{21}, \mu_{22})} 

and 

\eqn{\Sigma_2 = \left({\begin{array}{cc} \tau^2_{21} & \rho\tau_{21}\tau_{22} \\ \rho\tau_{21}\tau_{22} &  \tau^2_{22}\end{array}}\right)}.

The interaction paramter \eqn{\eta_{j}} is exchangeable with the other strata and follows a normal distribution,

\eqn{\eta_{j} = \sim N(\mu_{\eta}, \tau^2_{\eta})}.

With probability \eqn{1-p_j}, \eqn{\theta_{1j}}, \eqn{\theta_{2j}} and, \eqn{\eta_{j}} are nonexchangeable with other strata. The prior distributions for \eqn{\theta_{1j}} and \eqn{\theta_{2j}} follow a bivariate normal distribution with mean vector \eqn{m_{1W}} and covariance matrix \eqn{S_{1W}} for agent 1 and mean vector \eqn{m_{2W}} and covariance matrix \eqn{S_{2W}} for agent 2,

\eqn{\theta_{1j} \sim BVN(m_{1W}, S_{1W})},

\eqn{\theta_{2j} \sim BVN(m_{2W}, S_{2W})}.

The prior distribution for \eqn{\eta_{j}} follows a normal distribution with

\eqn{\eta_{j} \sim N(\mu_{\eta_{0}}, \tau^2_{\eta_{0}})}.


Details on deriving the priors can be found in the references below.
}
\examples{
\dontrun{
## Example 1: Analysis of three concurrent combination phase I trials with 2 agents

threeTrial2Agent<-Combo2BLRM.EXNEX(
 DoseRef1=16,
 Prior.NEX.Agent1=c(log(0.15/0.85),0,2,1,0),
 Prior.EX.mu1.Agent1=c(log(0.15/0.85),2),
 Prior.EX.mu2.Agent1=c(0,1),
 Prior.EX.tau1.Agent1=c( log(0.25), log(2)/1.96),
 Prior.EX.tau2.Agent1=c( log(0.125), log(2)/1.96),
 Prior.rho.Agent1=c(-1,1),
 DoseRef2=2,
 Prior.NEX.Agent2=c(log(0.1/0.9),0,2,1,0),
 Prior.EX.mu1.Agent2=c(log(0.1/0.9),2),
 Prior.EX.mu2.Agent2=c(0,1),
 Prior.EX.tau1.Agent2=c( log(0.25), log(2)/1.96),
 Prior.EX.tau2.Agent2=c( log(0.125), log(2)/1.96),
 Prior.rho.Agent2=c(-1,1),
 Prior.EX.eta.mu=c(0, 1.121),
 Prior.NEX.eta=c(0, 1.121),
 Prior.EX.eta.tau=c( log(0.25), log(2)/1.96),
 Pint=c(0.16,0.33),
 MCMC=c(2500, 7500, 4, 1),
 Rnd.seed=1452671,
 RSeed.WinBUGS=1367,
 DIC=TRUE,
 progress.bar="text",
 agent1="PF1",
 agent2="PF2",
 pj.EX=0.5,
 Doses1=c(2,4,8,16,36),
 Doses2=c(0,1,2),
 trialData=list(
   
  list(DosesAdm1=c(1,2,5,10,25,50),
        DosesAdm2=c(2,2,2,2,2,2),
        Npat=c(3,3,2,3,6,3),
        Ntox=c(0,0,0,0,0,2),
        trialName="trialAB1"),
   
   list(DosesAdm1=c(5,10,10,10,20),
        DosesAdm2=c(2,2,2.5,3,2),
        Npat=c(3,3,3,3,3),
        Ntox=c(0,0,0,1,0),
        trialName="TrialAB2"),
   
   
   list(DosesAdm1=c(1,2,4,8),
        DosesAdm2=c(0,0,0,0),
        Npat=c(3,3,6,3),
        Ntox=c(0,0,1,0),
        trialName="trialA")
 )
 
)                          
}


}
\references{
Neuenschwander B, Wandel S, Roychoudhury S, Bailey S. Robust exchangeability designs for early phase clinical trials with multiple strata. Pharm Stat. 2016;15(2):123-34.

Neuenschwander B, Roychoudhury S, Schmidli H. On the use of co-data in clinical trials. Biopharmaceutical Research. 2016;8(3):345-54

Neuenschwander B, Matano A, Tang Z, Roychoudhury S,Wandel S, Bailey S. A Bayesian Industry Approach to Phase I Combination Trials in Oncology. Statistical Methods for Drug Combination Studies. Taylor & Francis, 2015.
}
