% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combo2BLRM.R
\name{Combo2BLRM}
\alias{Combo2BLRM}
\title{Combo2BLRM Function}
\usage{
Combo2BLRM(
  Data = NULL,
  Agent1 = NULL,
  Doses1 = NULL,
  DoseRef1 = NULL,
  Prior.logAB.1 = NULL,
  pmix.logAB1 = c(1),
  Agent2 = NULL,
  Doses2 = NULL,
  DoseRef2 = NULL,
  Prior.logAB.2 = NULL,
  pmix.logAB2 = c(1),
  Prior.eta = NULL,
  pmix.eta = c(1),
  Prior.Only = FALSE,
  Pcutoffs = c(0.16, 0.33),
  New.cohortsize = 3,
  Outfile = "Out3",
  Plot = TRUE,
  int.crit = c(1, 1, 0.25),
  int.col = list("lightgreen", "green", c("yellow", "red")),
  pars = c("logAlphaBeta1", "logAlphaBeta2", "eta", "P12", "pCat", "pred.Tox"),
  MCMC = c(2500, 7500, 4, 1),
  Warn = TRUE,
  DIC = TRUE,
  progress.bar = "text",
  Rnd.seed = 1452671,
  RSeed.WinBUGS = 1367
)
}
\arguments{
\item{Data}{Dose DLT data -- a list objest contains DoseAdm1, DoseAdm2, Npat, Ntox, and Weight.
DoseAdm1 is a vector of first agent doses which are administered at patient cohorts. 
DoseAdm2 is a vector of second agent doses which are administered at patient cohorts. 
Npat is a vector of same length as DoseAdm, the elements are the cohort sizes at each DoseAdm. 
Ntox is a vector of same length as DoseAdm, the elements are the the number of toxicitites 
observed for each cohort. Weight is a vector of same length as DoseAdm, the elements are 1 
by default, but can be changed depending on how much information one wants borrow form this 
data set; 1 is maximum i.e. borrowing the full information, 0 is minimum i.e. borrowing 
no information, similary, setting 0.5 means borrowing half of the information}

\item{Agent1}{Compound 1 name}

\item{Doses1}{Doses for compound 1}

\item{DoseRef1}{Dose reference for compound 1}

\item{Prior.logAB.1}{Prior for single agent effect of compound 1 (a list object)}

\item{pmix.logAB1}{mixing probability of the prior distributions for compound 1. 
Usually is set to c(1) however this is a vector of mixing probabilities when 
mixture of priors is used. See the package vignettes for details}

\item{Agent2}{Compound 2 name}

\item{Doses2}{Doses for compound 2}

\item{DoseRef2}{Dose reference for compound 2}

\item{Prior.logAB.2}{Prior for single agent effect of compound 2 (a list object)}

\item{pmix.logAB2}{mixing probability of the prior distributions for compound 2. 
Usually is set to c(1) however this is a vector of mixing probabilities when 
mixture of priors is used. See the package vignettes for details}

\item{Prior.eta}{prior for interaction parameter \eqn{\eta}, a vector of mean and standard deviation}

\item{pmix.eta}{mixing probability of the prior distributions of the interaction parameter. 
Usually is set to c(1) however this is a vector of mixing probabilities when 
mixture of priors is used.}

\item{Prior.Only}{either TRUE or FALSE. If TRUE, Data must be equal to NULL meaning that there is 
no data and posterior summaries are basically the summaries from the prior distribution. If FALSE,
then Data must be provided and posterior will be computed.}

\item{New.cohortsize}{new for prediction}

\item{Outfile}{output file}

\item{Plot}{Indicator if Plot will be generated}

\item{int.crit}{Critical value for interval to change color}

\item{int.col}{Color of each bar}

\item{pars}{MCMC parameters to save}

\item{MCMC}{MCMC specifications -- nburn: burnin length, ntot: Markov chain length after burnin, 
nchain: number of Markov chains, nthin: thinning interval, default is 1 (no thinning)}

\item{Warn}{Prints convergence warning}

\item{DIC}{DIC to save}

\item{progress.bar}{Jags parameter for type of progress bar. Possible values are
"text", "gui", and "none". Type "text" is displayed on the R console. Type "gui"
is a graphical progress bar in a new window. The progress bar is suppressed if
progress.bar is "none"}

\item{Rnd.seed}{Seed for R}

\item{RSeed.WinBUGS}{Seed for WinBugs}
}
\value{
\item{Data}{The dataset}
\item{Priors}{Priors used for the analysis}
\item{logAlphaBeta1}{Posterior summary of log(\eqn{\alpha}) and log(\eqn{\beta}) for agent 1}
\item{logAlphaBeta.cor1}{Posterior summary of the correlation parameter for agent 1}
\item{logAlphaEBeta1}{Posterior summary of log(\eqn{\alpha}) and \eqn{\beta} for agent 1}
\item{logAlphaBeta2}{Posterior summary of log(\eqn{\alpha}) and log(\eqn{\beta}) for agent 2}
\item{logAlphaBeta.cor2}{Posterior summary of the correlation parameter for agent 2}
\item{logAlphaEBeta2}{Posterior summary of log(\eqn{\alpha}) and \eqn{\beta} for agent 2}
\item{eta}{Posterior summary of intercation paramter \eqn{\eta})}
\item{OddsFactor}{Posterior summary of the odds factor}
\item{P}{Predicted summaries of the dose levels}
\item{Pcat}{Posterior predictive probabilities of doses}
\item{PcatP}{Posterior predictive probabilities and summaries of doses}
\item{pred.Tox}{Predictive toxicity probilities of doses}
\item{Post.pmix1}{Posterior mixture weigths for agent 1}
\item{Post.pmix2}{Posterior mixture weigths for agent 2}
\item{Post.pmix2}{Posterior mixture weigths for \eqn{\eta}}
\item{Rhat}{The MCMC convergence dignostoc Rhat for all the parameters}
\item{MCMC}{MCMC informations}
\item{R2WB}{}
In addition, a text output containing all of the above and two plots -- posterior DLT rates and 
interval probabilities of DLT's, are produced.
}
\description{
This function performs analysis with dual combination model
}
\details{
See vignettes. browseVignettes("OncoPh1BLRM")
}
\examples{
\dontrun{
# Doses for agent 1
Doses1 <- c(0, 3, 4.5, 6)
DoseRef1 <- 3

#Doses for agent 2
Doses2 <- c(0, 400, 600, 800)
DoseRef2 <- 800

# Priors derived using MAP analysis
MAP.prior.Agent1 <- c(-2.53, 0.03, 1.02, 0.86, -0.54)
MAP.prior.Agent2 <- c(-2.12, 0.69, 0.61, 0.64, -0.20)

# Interaction prior
prior.interaction <- function(Odds.median= NULL, Odds.upper= NULL){
mean.eta <- log(Odds.median)
sd.eta <- (log(Odds.upper) - log(Odds.median))/1.96
return(list(mean.eta=mean.eta, sd.eta=sd.eta))
}

# Prior for eta: Median: No increase in Odds of DLT Upper: 9 fold increase in Odds of DLT
peta <- prior.interaction(Odds.median = 1.1, Odds.upper = 9)
etaPrior <- c(peta$mean.eta, peta$sd.eta)

# Duel agent data
data.combo2 <- list(DosesAdm1 = c(   3,   3,   6,   3,   6),
                    DosesAdm2 = c( 400, 800, 400, 800, 400),
                    Npat      = c(   3,   3,   3,   3,   3),
                    Ntox      = c(   0,   1,   1,   1,   0),
                    Weight    = c(   1,   1,   1,   1,   1)
)

# Analysis using dual agent BLRM
fit.BLRM.dual <-  Combo2BLRM(Data           = data.combo2 ,
                             Agent1         = "DRUG 1",
                             Doses1         = Doses1,
                             DoseRef1       = DoseRef1,
                             Prior.logAB.1  = list(MAP.prior.Agent1),
                             pmix.logAB1    = c(1),
                             Agent2         = "DRUG 2",
                             Doses2         = Doses2,
                             DoseRef2       = 800,
                             Prior.logAB.2  = list(MAP.prior.Agent2),
                             pmix.logAB2    = c(1),
                             Prior.eta      = list(etaPrior),
                             Prior.Only     = FALSE,
                             Pcutoffs       = c(0.16, 0.33),
                             New.cohortsize = 3,
                             Outfile        = "Output_dual",
                             MCMC           = c(2500, 15000, 4, 1),
                             DIC            = TRUE,
                             Rnd.seed       = 1452671,
                             RSeed.WinBUGS  = 1367
)
}



}
\references{
Neuenschwander B, Matano A, Tang Z, Roychoudhury S,Wandel S, Bailey S. 
A Bayesian Industry Approach to Phase I Combination Trials in Oncology. 
Statistical Methods for Drug Combination Studies. Taylor & Francis, 2015.
}
