% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combo3BLRM.R
\name{Combo3BLRM}
\alias{Combo3BLRM}
\title{Combo3BLRM Function}
\usage{
Combo3BLRM(
  Data = NULL,
  Agent1 = NULL,
  Doses1 = NULL,
  DoseRef1 = NULL,
  Prior.logAB.1 = NULL,
  pmix.logAB1 = c(1),
  Agent2 = NULL,
  Doses2 = NULL,
  DoseRef2 = NULL,
  Prior.logAB.2 = NULL,
  pmix.logAB2 = c(1),
  Agent3 = NULL,
  Doses3 = NULL,
  DoseRef3 = NULL,
  Prior.logAB.3 = NULL,
  pmix.logAB3 = c(1),
  Prior.eta = NULL,
  Prior.Only = FALSE,
  Pcutoffs = c(0.16, 0.33),
  New.cohortsize = 3,
  Outfile = "Out3",
  Plot = TRUE,
  int.crit = c(1, 1, 0.25),
  int.col = list("lightgreen", "green", c("yellow", "red")),
  pars = c("logAlphaBeta1", "logAlphaBeta2", "logAlphaBeta3", "eta12", "eta13",
    "eta23", "eta123", "P12", "pCat", "pred.Tox"),
  MCMC = c(2500, 7500, 4, 1),
  Warn = TRUE,
  DIC = TRUE,
  progress.bar = progress.bar,
  Rnd.seed = 1452671,
  RSeed.WinBUGS = 1367
)
}
\arguments{
\item{Data}{Dose DLT data -- a list objest contains DoseAdm1, DoseAdm2, Npat, Ntox, and Weight.
DoseAdm1 is a vector of first agent doses which are administered at patient cohorts. 
DoseAdm2 is a vector of second agent doses which are administered at patient cohorts. 
DoseAdm3 is a vector of third agent doses which are administered at patient cohorts. 
Npat is a vector of same length as DoseAdm, the elements are the cohort sizes at each DoseAdm. 
Ntox is a vector of same length as DoseAdm, the elements are the the number of toxicitites 
observed for each cohort. Weight is a vector of same length as DoseAdm, the elements are 1 
by default, but can be changed depending on how much information one wants borrow form this 
data set; 1 is maximum i.e. borrowing the full information, 0 is minimum i.e. borrowing 
no information, similary, setting 0.5 means borrowing half of the information}

\item{Agent1}{Compound 1}

\item{Doses1}{Doses for compound 1}

\item{DoseRef1}{Reference dose for compound 1}

\item{Prior.logAB.1}{Prior for single agent effect of compound 1 (a list object)}

\item{pmix.logAB1}{mix probability compound 1}

\item{Agent2}{Compound 2}

\item{Doses2}{Doses for Compound 2}

\item{DoseRef2}{Dose reference for compound 2}

\item{Prior.logAB.2}{Prior for single agent effect of compound 2 (a list object)}

\item{pmix.logAB2}{mix probability compound 2}

\item{Agent3}{Compound 3}

\item{Doses3}{Doses for compound 3}

\item{DoseRef3}{Reference dose for compound 3}

\item{Prior.logAB.3}{Prior for single agent effect of compound 3 (a list object)}

\item{pmix.logAB3}{mix probability compound 3}

\item{Prior.eta}{Prior for interaction parameters (a list object with \eqn{\eta_{12}}, \eqn{\eta_{13}},
\eqn{\eta_{23}}, and \eqn{\eta_{123}})}

\item{Prior.Only}{PriorOnly}

\item{New.cohortsize}{new for prediction}

\item{Outfile}{output file}

\item{Plot}{Indicator if Plot will be generated}

\item{int.crit}{Critical value for interval to change color}

\item{int.col}{Color of each bar}

\item{pars}{MCMC parameters to save}

\item{MCMC}{MCMC specifications (nburn, ntot, nchain, nthin)}

\item{Warn}{Prints convergence warning}

\item{DIC}{DIC to save}

\item{progress.bar}{Jags parameter for type of progress bar. Possible values are
“text”, “gui”, and “none”. Type “text” is displayed on the R console. Type “gui”
is a graphical progress bar in a new window. The progress bar is suppressed if
progress.bar is “none”}

\item{Rnd.seed}{seed for R}

\item{RSeed.WinBUGS}{seed for WinBugs}
}
\value{
\item{Data}{The dataset}
\item{Priors}{Priors used for the analysis}
\item{logAlphaBeta1}{Posterior summary of log(\eqn{\alpha}) and log(\eqn{\beta}) for agent 1}
\item{logAlphaBeta.cor1}{Posterior summary of the correlation parameter for agent 1}
\item{logAlphaEBeta1}{Posterior summary of log(\eqn{\alpha}) and \eqn{\beta} for agent 1}
\item{logAlphaBeta2}{Posterior summary of log(\eqn{\alpha}) and log(\eqn{\beta}) for agent 2}
\item{logAlphaBeta.cor2}{Posterior summary of the correlation parameter for agent 2}
\item{logAlphaEBeta2}{Posterior summary of log(\eqn{\alpha}) and \eqn{\beta} for agent 2}
\item{logAlphaBeta3}{Posterior summary of log(\eqn{\alpha}) and log(\eqn{\beta}) for agent 3}
\item{logAlphaBeta.cor3}{Posterior summary of the correlation parameter for agent 3}
\item{logAlphaEBeta3}{Posterior summary of log(\eqn{\alpha}) and \eqn{\beta} for agent 3}
\item{eta12}{Posterior summary of intercation paramter \eqn{\eta_{12}})}
\item{eta23}{Posterior summary of intercation paramter \eqn{\eta_{23}})}
\item{eta13}{Posterior summary of intercation paramter \eqn{\eta_{13}})}
\item{eta123}{Posterior summary of intercation paramter \eqn{\eta_{123}})}
\item{OddsFactor12}{Posterior summary of the odds factor between first and second agent}
\item{OddsFactor23}{Posterior summary of the odds factor between second and third agent}
\item{OddsFactor13}{Posterior summary of the odds factor between first and third agent}
\item{OddsFactor123}{Posterior summary of the odds factor between first, second, and third agent}
\item{P}{Predicted summaries of the dose levels}
\item{Pcat}{Posterior predictive probabilities of doses}
\item{PcatP}{Posterior predictive probabilities and summaries of doses}
\item{pred.Tox}{Predictive toxicity probilities of doses}
\item{Post.pmix1}{Posterior mixture weigths for agent 1}
\item{Post.pmix2}{Posterior mixture weigths for agent 2}
\item{Post.pmix3}{Posterior mixture weigths for agent 3}
\item{Rhat}{The MCMC convergence dignostoc Rhat for all the parameters}
\item{MCMC}{MCMC informations}
\item{R2WB}{}
In addition, a text output containing all of the above and two plots -- posterior DLT rates and 
interval probabilities of DLT's, are produced.
}
\description{
This function performs analysis with triple combination model
}
\details{
See vignettes. browseVignettes("OncoPh1BLRM")
}
\examples{
 
\dontrun{


# The following example is a dose escalation study 
# where the escalation is done with respect to the first agent doses
# The study start at the maximum tolerated dose (MTD) found by the dose escalation study of the first agent
# along with the prespecified fixed doses of the other two agents

# Form Interaction prior
prior.interaction <- function(Odds.median = NULL, Odds.upper = NULL){
 mean.eta <- log(Odds.median)
 sd.eta <- (log(Odds.upper) - log(Odds.median))/1.96
 return(list(mean.eta=mean.eta, sd.eta=sd.eta))
}

# Prior for eta: Median -- No increase in Odds of DLT, Upper -- 9 fold increase in Odds of DLT
peta <- prior.interaction(Odds.median = 1, Odds.upper = 9)
etaPrior <- c(peta$mean.eta, peta$sd.eta)
etaPrior

eta.prior = list(etaPrior12  = etaPrior,
                 etaPrior13  = etaPrior,
                 etaPrior23  = etaPrior,
                 etaPrior123 = etaPrior
)



Doses.Agent1 <- c(8, 16, 32, 64, 128, 256, 512)  # 8 is starting dose, 50 is an intermediate dose to be investigated
# 500 is not tolerable
DoseRef.Agent1 <- 275  # Expected MTD 50 + (50 + 500)/2
MTD.Agent1     <- 128   # However, this is delcared MTD after the dose escalation experiment


require(boot)
weakly.informative.prior <- c(logit(0.33), 0, 2, 1, 0)  # Default


# Doses.Agent2   <- c(0.02, 1, 2.5, 5, 10, 30)  # Used for monotherpy
DoseRef.Agent2 <- 2.5

# Doses.Agent3   <- c(25, 50, 75, 100, 125, 150)  # Used for monotherpy
DoseRef.Agent3 <- 125

Doses.Agent1 <- c(Doses.Agent1, MTD.KaT6)  
# For this agent the data are dose escalation data and the MTD
Doses.Agent2 <- c(rep(0, length(Doses.Agent1)), DoseRef.Letro) 
# For this agent the data are 0's and the prespecified fixed dose
Doses.Agent3 <- c(rep(0, length(Doses.Agent1)), DoseRef.Palbo) 
# For this agent the data are 0's and the prespecified fixed dose


# Triple agent data
Data.combo3 <- list(DosesAdm1 = MTD.Agent1,
                    DosesAdm2 = DoseRef.Agent2,
                    DosesAdm3 = DoseRef.Agent3,
                    Npat      = c(6),
                    Ntox      = c(0),
                    Weight    = c(1)
)



fit.triple <- Combo3BLRM(Data           = Data.combo3, 
                         Agent1         = "Agent1", 
                         Doses1         = Doses.Agent1,
                         DoseRef1       = DoseRef.Agent1, 
                         Prior.logAB.1  = list(weakly.informative.prior), 
                         pmix.logAB1    = c(1),
                         Agent2         = "Agent2", 
                         Doses2         = Doses.Agent2, 
                         DoseRef2       = DoseRef.Agent2,
                         Prior.logAB.2  = list(weakly.informative.prior), 
                         pmix.logAB2    = c(1), 
                         Agent3         = "Agent3",
                         Doses3         = Doses.Agent3, 
                         DoseRef3       = DoseRef.Agent3, 
                         Prior.logAB.3  = list(weakly.informative.prior),
                         pmix.logAB3    = c(1), 
                         Prior.eta      = eta.prior, 
                         Prior.Only     = FALSE,
                         Pcutoffs       = c(0.16, 0.33), 
                         New.cohortsize = 3, 
                         Outfile        = "Triple_0_6_128",
                         Plot           = TRUE, 
                         int.crit       = c(1, 1, 0.25), 
                         int.col        = list("lightgreen", "green", c("yellow", "red")), 
                         pars           = c("logAlphaBeta1", "logAlphaBeta2", "logAlphaBeta3", 
                                            "eta12", "eta13", "eta23", "eta123", "P12", "pCat", "pred.Tox"), 
                         MCMC           = c(20000, 30000, 4, 1), 
                         Warn           = TRUE,
                         DIC            = TRUE, 
                         Rnd.seed       = 1452671,
                         RSeed.WinBUGS  = 1367)

fit.triple$PcatP[,, 2, 2, ]

}


}
\references{
Neuenschwander B, Matano A, Tang Z, Roychoudhury S,Wandel S, Bailey S. A Bayesian Industry
Approach to Phase I Combination Trials in Oncology. Statistical Methods for Drug Combination
Studies. Taylor & Francis, 2015.
}
