% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_BLRM.R
\name{MAP.BLRM}
\alias{MAP.BLRM}
\title{Meta analytic Predictive Prior}
\usage{
MAP.BLRM(
  Agent = NULL,
  Doses = NULL,
  DoseRef = NULL,
  Data = NULL,
  Prior.muA = c(0, 2),
  Prior.muB = c(0, 1),
  Nexch = 1,
  tau.index = NULL,
  Prior.tauA = NULL,
  Prior.tauB = NULL,
  Prior.rho = c(-1, 1),
  Pcutoffs = c(0.16, 0.33),
  New.cohortsize = 3,
  Outfile = "Out",
  pars = c("mu.ex", "logAB", "pTox", "pCat", "MAP.pTox", "MAP.pCat", "pred.Tox",
    "MAP.logAB", "tau", "rho", "MAP.logAB.XY", "logAB.XY"),
  MCMC = c(2500, 7500, 4, 1),
  Warn = TRUE,
  DIC = TRUE,
  progress.bar = "text",
  Rnd.seed = 1452671,
  RSeed.WinBUGS = 1367
)
}
\arguments{
\item{Agent}{name of the Agent under investigation.}

\item{Doses}{a vector of Doses for which inferential results will be provided.}

\item{DoseRef}{refernce dose for MAP.BLRM.}

\item{Data}{external dose-DLT data from different strata: a list object contains DoseAdm, Npat, Ntox, Study and StudyLabel.
DoseAdm, Npat, and Ntox represent vectors of doses administered, Npat number of patients treated and the number of
dose-limiting toxicitites observed in different strata. Study is a vector indicating different strata. Here strata represents
study, subgroups and other source of external data. The lenght of DoseAdm, Npat, Ntox, and Study must be same. StudyLabel is 
a vector of strata names. The length of  StudyLabel needs to be unique number of elements in Study.}

\item{Prior.muA}{prior mean and sd of the normal distribution for the mean \eqn{\mu_{\log(\alpha)}} 
of the exchangeability distribution (see details).}

\item{Prior.muB}{prior mean and sd of the normal distribution for the mean \eqn{\mu_{\log(\beta)}} 
of the exchangeability distribution (see details).}

\item{Nexch}{Number of exchangeable set for between-strata standard deviations. Default value is 1 (full exchangeability model).}

\item{tau.index}{\eqn{(Nstrata+1)} vector with indices for between-strata standard deviations; indices
must be in \eqn{1.\ldots ,Nexch}. tau.index = 1 suffices for the full exchangeability model.}

\item{Prior.tauA}{\eqn{Nexch \times 2} matrix with mean (1st column) and sd (2nd column) 
of the log-normal prior(s) for \eqn{\tau_{\log(\alpha)}} of the exchangeability distribution; 
\eqn{Nexch} is the number of between-strata standard deviations (see tau.index and details).}

\item{Prior.tauB}{\eqn{Nexch \times 2} matrix with mean (1st column) and sd (2nd column) 
of the log-normal prior(s) for \eqn{\tau_{\log(\beta)}} of the exchangeability distribution; 
\eqn{Nexch} is the number of between-strata standard deviations (see tau.index and details).}

\item{Prior.rho}{vector with parameters of uniform distribution for correlation \eqn{\rho} of the exchangeability
distribution of \eqn{(\log(\alpha), \log(\beta))}(see details). The default is Uniform(-1,1).}

\item{Pcutoffs}{cutoffs to define the intervals for the true DLT probability. The default is
(0.16, 0.33) defining three categories: underdosing ({[}0--0.16{]}), target ((0.16--0.33)), and 
overdosing ({[}0.33--1.00{]}).}

\item{New.cohortsize}{number of subjects in a new cohort, for which the predictive distribution 
of the number of DLT's will be computed.}

\item{Outfile}{name of the output file.}

\item{pars}{MCMC parameters to save for output summary.}

\item{MCMC}{four component vector of MCMC parameters n.burnin, n.iterations, n.chains, n.thin.
See \link{R2WinBUGS} for details.}

\item{Warn}{logical: if TRUE, the convergence warning will be printed.}

\item{DIC}{logical: if TRUE, compute deviance, pD, and DIC. See \link{R2WinBUGS} for details.}

\item{progress.bar}{progress.bar type of progress bar. Possible values are “text”, “gui”, and “none”. Type “text”
is displayed on the R console. See \link{rjags} for details.}

\item{Rnd.seed}{Random seed for the R random stream. This is important for reproducibility.}

\item{RSeed.WinBUGS}{Random seed for the WinBUGS random stream. This is important for reproducibility.}
}
\value{
The function returns a list of the following main items:
\item{Data}{The dataset analyzed}
\item{Priors}{summary of prior specifications}
\item{MAP.logAB.sim}{MCMC sample of the predictive distribution of \eqn{(\log(\alpha^*),\log(\beta^*))}}
\item{MAP.Prior}{Bivariate normal approximation of MAP prior for \eqn{(\log(\alpha^*),\log(\beta^*))}}
\item{MAP.P}{summary of MAP prior for DLT probability in a new stratum}
\item{MAP.Pcat}{summary of interval probabilities for DLT probability in a new stratum}
\item{MAP.PcatP}{MAP.Pcat and MAP.P combined}
\item{logAB}{Posterior summaries BLRM parameters for all strata including new stratum (MAP.prior)}
\item{logAB.cor}{Estimated correlation for all strata including new stratum (MAP prior)}
\item{P}{summary of DLT probabilities for all strata including the new stratum (MAP prior)}
\item{Pcat}{summary of interval probabilities for DLT probability for all strata including the new stratum (MAP prior)}
\item{pred.Tox}{predicted probabilities for the number of DLT's for a cohort of size New.cohortSize for all strata including new stratum (MAP prior)}
\item{mu.tau.rho}{Posterior summary of the exchageability distribution and between-trial standard deviation}
\item{Rhat}{Gelman-Rubin MCMC convergence diagnostics}
\item{MCMC}{MCMC specifications}
\item{R2WB}{WinBUGS object with summaries of MCMC analysis (see \link{R2WinBUGS})}
In addition, a text output file (Outfile.txt) with all summaries is produced.
}
\description{
This function derives Meta-analytic predictive prior for single agent Bayesian logstic linear model (BLRM) parameters for Phase I Oncology trials.
}
\details{
Let \eqn{r_{d,s}} and \eqn{n_{d,s}} be the number of patients treated and the number of DLT's observed 
at dose level \eqn{d} from strata \eqn{s} .The aim is to derive a prior distribution 
for the logistic parameters of new stratum \eqn{(\log(\alpha^*),\log(\beta^*))}. The model specifications are 
as follows:
\deqn{r_{d,s} \sim Bin(\pi_{d,s},n_{d,s}); \; s=1, \ldots H} 
\deqn{\mbox{logit}(\pi_{d,s}) = \log(\alpha_s) + \beta_s \log(d/d^*)}
\deqn{\mbox{logit}({\pi_d}^*) = \log(\alpha^*) + \beta^* \log(d/d^*)}
\deqn{(\log(\alpha_s),\log(\beta_s))  \sim  BVN((\mu_{\log(\alpha)}, \mu_{\log(\beta)}), \Psi)}
\deqn{(\log(\alpha^*),\log(\beta^*))  \sim  BVN((\mu_{\log(\alpha)}, \mu_{\log(\beta)}), \Psi)}

where \eqn{\Psi}	is the between-strata covariance matrix with standard deviations \eqn{\tau_{\log(\alpha)}},
\eqn{\tau_{\log(\beta)}} and correlation \eqn{\rho}. The MAP prior of new stratum \eqn{(\log(\alpha^*),\log(\beta^*))}
is the predictive distribution and approximated by bivariate normal distribution:
\deqn{(\log(\alpha^*),\log(\beta^*))|(r_{d,s}, n_{d,s} : s= 1, \ldots, H)}

Furthermore, following prior distributions are used for model parameters:
 \itemize{
  \item normal priors are used for \eqn{\mu_{\log(\alpha)}} and \eqn{\mu_{\log(\beta)}}. 
  The expected toxicity at DoseRef can be corroborated by setting the mean appropriately.
  For example, when expected toxicity os 0.5 then the mean is log(0.5/(1 - 0.5)) which
  is default. 
  \item log-normal priors for \eqn{\tau_{\log(\alpha)}} and  \eqn{\tau_{\log(\beta)}} and 
  \item a uniform prior for \eqn{\rho}. 
}

For differential discounting, different \eqn{\tau} can be used (see tau.index). For typical
values for the medians of the log-normal prior distributions, see Neuenschwander 2015.
}
\examples{
\dontrun{

##Example 1: Prior derivation for Japanese study using Western data (example from 
section 6.3.3.3 in Neuenschwander 2015)

# Western first-time-in-human trial is specified as the historical data
Data.western <- list(DosesAdm    = c(1, 2, 4, 8, 15, 30),
                     Npat       = c(5, 6, 5, 9,  8,  4),
                     Ntox       = c(0, 0, 0, 0,  1,  3),
                     Study      = c(1, 1, 1, 1,  1,  1),
                     StudyLabel = c("Western")
                     )
                     
# MAP Prior for Japanese Phase I trial
prior.japan <- MAP.BLRM(Doses = c(1, 2, 4, 8, 12, 15, 30, 40),
               DoseRef        = 10,
               Data           = Data.western,
               Prior.muA      = c(0,2),
               Prior.muB      = c(0,1),
               Prior.tauA     = c( log(0.25), log(2)/1.96),
               Prior.tauB     = c( log(0.125), log(2)/1.96),
               Pcutoffs       = c(0.16, 0.33),
               New.cohortsize = 3,
               Outfile        = "Output_Japan_MAP",
               MCMC           = c(2500, 7500, 4, 1),
               DIC            = TRUE,
               Rnd.seed       = 1452671,
               RSeed.WinBUGS  = 1367
               )
               
MAP.prior.japan <- prior.japan$MAP.Prior
MAP.prior.japan

##Example 2: Differential discounting

}


}
\references{
Neuenschwander B, Matano A, Tang Z, Roychoudhury S,Wandel S, Bailey S. A Bayesian Industry
Approach to Phase I Combination Trials in Oncology. Statistical Methods for Drug Combination
Studies. Taylor & Francis, 2015.
}
