% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleBLRM_EXNEX.R
\name{SingleBLRM.EXNEX}
\alias{SingleBLRM.EXNEX}
\title{SingleBLRM.EXNEX Function}
\usage{
SingleBLRM.EXNEX(
  DoseRef = NULL,
  Prior.NEX = NULL,
  Prior.EX.mu1 = NULL,
  Prior.EX.mu2 = NULL,
  Prior.EX.tau1 = NULL,
  Prior.EX.tau2 = NULL,
  Prior.rho = c(-1, 1),
  Pint = c(0.16, 0.33),
  MCMC = c(2500, 7500, 4, 1),
  Rnd.seed = 1452671,
  RSeed.WinBUGS = 1367,
  DIC = TRUE,
  progress.bar = "text",
  pj.EX = NULL,
  Doses = NULL,
  trialData = NULL
)
}
\arguments{
\item{DoseRef}{refernce dose for the BLRM.}

\item{Prior.NEX}{is the nonexchangeability prior for the BLRM parameters, \eqn{log(\alpha)} and \eqn{log(\beta)}. This is a vector of length 5 for the NEX bivariate normal prior for, (\eqn{m_{log(\alpha)}}, \eqn{m_{log(\beta)}}, \eqn{s_{log(\alpha)}}, \eqn{s_{log(\beta)}}, \eqn{Corr(log(\alpha),log(\beta))}).}

\item{Prior.EX.mu1}{is the EX prior specification for the \eqn{log(\alpha)} mean parameter. It is a vector of length 2, (\eqn{m_{\mu1}, s_{\mu1}}).}

\item{Prior.EX.mu2}{is the EX prior specification for the \eqn{log(\beta)} mean parameter. It is a vector of length 2, (\eqn{m_{\mu2}, s_{\mu2}}).}

\item{Prior.EX.tau1}{is the EX prior specification for the \eqn{log(\alpha)} standard deviation parameter. It is a vector of length 2, (\eqn{m_{\tau1}, s_{\tau1}}).}

\item{Prior.EX.tau2}{is the EX prior specification for the \eqn{log(\beta)} standard deviation parameter. It is a vector of length 2, (\eqn{m_{\tau2}, s_{\tau2}}).}

\item{Prior.rho}{is the uniform prior specification for the correlation between \eqn{\tau1} and \eqn{\tau2}. It is a vector of length 2, and the default is set to (-1,1).}

\item{Pint}{targeted interval probability cutoff, which will be used to define three intervals for the true DLT probability. The default is
(0.16, 0.33) defining three categories: underdosing ({[}0--0.16{]}), target ((0.16--0.33)), and 
overdosing ({[}0.33--1.00{]}).}

\item{MCMC}{four component vector of MCMC parameters n.burnin, n.iterations, n.chains, n.thin. See \link{R2WinBUGS} for details.}

\item{Rnd.seed}{Random seed for the R random stream. This is important for reproducibility.}

\item{RSeed.WinBUGS}{Random seed for the WinBUGS random stream. This is important for reproducibility.}

\item{DIC}{logical: if TRUE, compute deviance, pD, and DIC. See \link{R2WinBUGS} for details.}

\item{progress.bar}{type of progress bar. Possible values are "text", "gui", and "none". Type "text"
is displayed on the R console. See \link{rjags} for details.}

\item{pj.EX}{is the proportion of exchangeability, resulting in a 1- pj.EX nonexchangeability proportion. Currenlty set to be the same for all strata. For example, pj.EX = 1 is full exchangeability, while pj.EX=0 is a stratified analysis.}

\item{Doses}{a vector of Doses for which inferential results will be provided, i.e., the provisional dose levels.}

\item{trialData}{dose-DLT data: a list with length equal to the number of trials. Data for each trial is entered as a list object with 4 elements; DoseAdm, Npat, Ntox, trialName. DoseAdm is a vector of the doses administered. Npat are the number of participants treated at a given dose. Ntox are the number of DLTs observed for a given dose. trialName is the unique name used to identfy a specific trial.}
}
\value{
The function returns as a list the following main items:
\item{P}{Posterior summary of DLT probabilities}
\item{Pcat}{Posterior summary of DLT interval probabilities}
\item{PcatP}{P and Pcat combined}
\item{logAlphaBeta}{Posterior summary of (\eqn{\log(\alpha)}, \eqn{\log(\beta)})}
\item{PmixEXNEX}{Posterior mixture weigths}
\item{intervalPlot}{Interval plot for DLT rates}
\item{postProbPlot}{Estimates and 95\% interval plots for DLT rates}
\item{samplingPlot}{Summary of the sampling results including the interval DLT rates}
}
\description{
This function analyzes dose limiting toxicity (DLT) data using a Bayesian logistic regression model (BLRM) combined with the Exchangeability-nonexchangeability (EXNEX) approach for single agent Phase I Oncology trials. The EXNEX approach allows information to be borrowed across concurrent trials.
}
\details{
Statistical model:

EXNEX for binary data:
The EXNEX apporach allows the use of co-data where stratum specific (e.g. study specific) parameters can be either exchangeable with parameters from other strata or nonexchangeable. For binary data this is the log-odds parameter, \eqn{\theta_j = log(\frac{\pi_j}{1-\pi_j})}.

For each strata, j, we have weights \eqn{p_j} and \eqn{1-p_j}. Where the parameters are exchangeable (EX) with probability \eqn{p_j} and \eqn{\theta_j} follows a normal distribution with exchangeability parameters \eqn{\mu} and \eqn{\tau},

\eqn{\theta_j \vert \mu, \tau \sim N(\mu,\tau^2)}

and nonexchangeable (NEX) with probability \eqn{1-p_j} and strata-specific priors are used where,

\eqn{\theta_j \sim N(m_j,v_j)}.

BLRM and EXNEX:

The dose-toxicity relationship will be modeled using a 2-parameter BLRM where for dose \eqn{d^*}, the number of participants with a dose limiting toxicity (DLT), \eqn{r_d}, in cohort of size \eqn{n_d} is binomial,

\eqn{r_d \vert \pi_d \sim Binomial(\pi_d, n_d)}

and for each stratum, j, the dose-toxicity model is logistic,

\eqn{logit(\pi_{jd}) = log(\alpha_j) + \beta_j log(\frac{d}{d^*})}

where \eqn{d} is the current dose under evaluation and \eqn{d^*} is the reference dose, which is used to scale the doses. The \eqn{\alpha, \beta > 0} are the parameters of the model such that \eqn{\alpha} is the therapy's odds of a DLT at \eqn{d^*}, and \eqn{\beta} is the increase in the log-odds of a DLT for a unit increase in the log-dose.

The model is then confined to one EX distribution and one NEX distribution, with *a priori* \eqn{p_j} and \eqn{1-p_j}. Under \eqn{p_j = 1} parameters are fully exchangeable, while \eqn{p_j = 0} would result in a fully stratified analysis with no exchangeability.

With probability \eqn{p_j} for EX, the logistic parameter vector

\eqn{\theta_j = [log(\alpha_j),log(\beta_j)]}

is exchangeable with the other strata parameters and follows a bivaraite normal distribution with mean vector \eqn{\mu} and covariance matrix \eqn{\Sigma};

\eqn{\theta_j \vert \mu, \Sigma \sim BVN(\mu,\Sigma)}

where

\eqn{\mu = (\mu_1, \mu_2)} 

and 

\eqn{\Sigma = \left({\begin{array}{cc} \tau^2_1 & \rho\tau_1\tau_2 \\ \rho\tau_1\tau_2 &  \tau^2_2\end{array}}\right)}.

With probability \eqn{1-p_j}, \eqn{\theta_j} is nonexchangeable with other strata and the prior distribution follows a bivariate normal distribution with mean vector \eqn{m_W} and covariance matrix \eqn{S_W},

\eqn{\theta_j \sim BVN(m_W, S_W)}.

Prior Specification:
The Bayesian approach requires the specification of the prior distributions for the model parameters \eqn{log(\alpha)} and \eqn{log(\beta)}. Below are the prior detials for both EX and NEX.

Details on prior derivation can be found in the references below.
}
\examples{
\dontrun{
## Example 1: Analysis of three concurrent phase I single agent trials 

threeSingleAgent<-SingleBLRM.EXNEX(
 DoseRef=24,
 Prior.NEX=c(log(0.13),0,2,1,0),
 Prior.EX.mu1=c(log(0.13), 1.982),
 Prior.EX.mu2=c(0, 0.992),
 Prior.EX.tau1=c(log(0.25), 0.3542),
 Prior.EX.tau2=c(log(0.125), 0.3542),
 Prior.rho=c(-1,1),
 MCMC=c(2500, 7500, 4, 1),
 Rnd.seed=1452671,
 RSeed.WinBUGS=1367,
 DIC=TRUE,
 progress.bar="text",
 pj.EX=0.5,
 Doses=c(4, 6, 12, 18, 24),
 trialData=list(

   list(DosesAdm=c(6,12,24),
        Npat=c(3,3,3),
       Ntox=c(0,0,0),
        trialName="Trial A"),

   list(DosesAdm=c(6,12,24),
        Npat=c(3,3,3),
        Ntox=c(0,1,0),
        trialName="Trial B"),

   list(DosesAdm=c(6,10,11,15),
        Npat=c(3,3,2,3),
        Ntox=c(0,1,0,0),
        trialName="Trial C")

 )
)                          
}


}
\references{
Neuenschwander B, Wandel S, Roychoudhury S, Bailey S. Robust exchangeability designs for early phase clinical trials with multiple strata. Pharm Stat. 2016;15(2):123-34.

Neuenschwander B, Roychoudhury S, Schmidli H. On the use of co-data in clinical trials. Biopharmaceutical Research. 2016;8(3):345-54
}
