% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleBLRM_PK.R
\name{SingleBLRM.PK}
\alias{SingleBLRM.PK}
\title{SingleBLRM.PK Function}
\usage{
SingleBLRM.PK(
  Data = NULL,
  Agent = NULL,
  Doses = NULL,
  Doses.Label = NULL,
  DoseRef = NULL,
  Prior.logAB = NULL,
  pmix.logAB = c(1),
  monotone.logAB = TRUE,
  Prior.logAB.PK = NULL,
  pmix.logAB.PK = c(1),
  monotone.PK = TRUE,
  Prior.sd.PK = NULL,
  random.model = TRUE,
  random.slope = FALSE,
  Prior.re.sd.PK = NULL,
  Prior.re.rho = c(0, 1),
  Xout = NULL,
  Xout.label = NULL,
  Prior.beta = c(0, 10),
  Prior.Only = FALSE,
  Pcutoffs = c(0.16, 0.33),
  New.cohortsize = 3,
  Outfile = "Out",
  Plot = TRUE,
  int.crit = c(1, 1, 0.25),
  pars = c("logAlphaBeta", "logGamma", "pTox", "pCat", "pred.Tox", "Zcat", "Zcat.PK",
    "sd.PK"),
  MCMC = c(2500, 7500, 4, 1),
  Warn = TRUE,
  DIC = TRUE,
  progress.bar = "text",
  Rnd.seed = 1452671,
  RSeed.WinBUGS = 1367
)
}
\arguments{
\item{Data}{Dose DLT data (a list objest contains DoseAdm,  Npat, Ntox, Weight and X (covariate for exposure))}

\item{Agent}{Compound name}

\item{Doses}{Doses for compound}

\item{Doses.Label}{Label for dose groups}

\item{DoseRef}{Dose reference for compound}

\item{Prior.logAB}{Prior for Bayesian Logistic Regession Model for Exposure-DLT (a list object)}

\item{pmix.logAB}{mix probability  Bayesian Logistic Regession Model for Exposure-DLT}

\item{monotone.logAB}{Monotonicity of Exposure-DLT model(values TRUE (default)/FALSE)}

\item{Prior.logAB.PK}{Prior for Dose-exposure model (a list object)}

\item{pmix.logAB.PK}{mix probability for Dose-exposure model}

\item{monotone.PK}{Monotonicity of Dose-exposure model(values TRUE (default)/FALSE)}

\item{Prior.sd.PK}{Prior for standard deviation for PK parameter}

\item{random.model}{logical: if TRUE, a random effect model will be used for dose-exposure analysis}

\item{random.slope}{logical:if TRUE, model with random intercept and slope will be used for dose-exposure model}

\item{Prior.re.sd.PK}{Prior for subject random effect standard deviation}

\item{Prior.re.rho}{prior for correlation between slope and interect of random effect}

\item{Xout}{Output pattern for covariates}

\item{Xout.label}{Output pattern labels}

\item{Prior.beta}{Prior for regression coefficient for covariates in Dose-exposure model}

\item{Prior.Only}{Prior only}

\item{New.cohortsize}{new for prediction}

\item{Outfile}{output file}

\item{Plot}{Indicator if Plot will be generated}

\item{int.crit}{Critical value for interval to change color}

\item{pars}{MCMC parameters to save}

\item{MCMC}{MCMC specifications (nburn, ntot, nchain, nthin)}

\item{Warn}{Prints convergence warning}

\item{DIC}{DIC to save}

\item{progress.bar}{Jags parameter for type of progress bar. Possible values are
type of progress bar. Possible values are "text", "gui", and "none". Type "text" is
displayed on the R console. Type "gui" is a graphical progress bar in a new window.
The progress bar is suppressed if progress.bar is "none"}

\item{Rnd.seed}{Seed for R}

\item{RSeed.WinBUGS}{Seed for WinBugs}

\item{int.col}{Color of each bar}
}
\value{
\item{Data}{Two list objects of doses and PK (exposure) inputed respectively}
\item{Priors}{Input priors}
\item{logAlphaBeta}{Posterior summary of (\eqn{\log(\alpha)}, \eqn{\log(\beta)})}
\item{logAlphaBeta.cor}{Posterior summary of the correlation between (\eqn{\log(\alpha)}, \eqn{\log(\beta)})}
\item{logAlphaEBeta}{Posterior summary of (\eqn{\log(\alpha)}, \eqn{\beta})}
\item{logGamma}{Posterior summary of (\eqn{\log(\gamma_0)}, \eqn{\log(\gamma_1)})}
\item{logGamma.cor}{Posterior summary of the correlation between (\eqn{\log(\gamma_0)}, \eqn{\log(\gamma_1)})}
\item{logEGamma}{Posterior summary of (\eqn{\log(\gamma_0)}, \eqn{\gamma_1})}
\item{sd.PK}{Posterior summary of subject random effect standard deviation}
\item{sd.re.PK}{Posterior summary of subject random effect standard deviation}
\item{beta}{Posterior summary of regression coefficient for covariates in Dose-exposure model}
\item{P}{Posterior summary of DLT probabilities}
\item{Pcat}{Posterior summary of DLT interval probabilities}
\item{PcatP}{P and Pcat combined}
\item{pred.Tox}{Predicted probabilities for the number of DLT's for a cohort of size New.cohortSize}
\item{Post.pmix}{Posterior mixture weigths (for mixture prior) for exposure-DLT model}
\item{Post.pmix.PK}{Posterior mixture weigths (for mixture prior) for Dose-exposure model}
\item{Rhat}{Gelman-Rubin MCMC convergence diagnostics}
\item{MCMC}{MCMC specifications}
\item{R2WB}{WinBUGS object with summaries of MCMC analysis (see \link{R2WinBUGS})}
}
\description{
This function performs analysis with Bayesian logistic regression (BLRM) with PK data for 
             single agent dose escalation
}
\details{
To be added
}
\examples{

\dontrun{


Data.Dose.PK <- list(DosesAdm = c(0.1, 0.1, 0.3, 0.3, 0.3, 1.0, 1.0, 3.0, 3.0, 3.0,
                                  10.0, 10.0, 30.0, 30.0, 30.0, 30.0, 30.0, 50.0, 50.0, 50.0),
                     Dose.Grp = c(1, 1, 2, 2, 2, 3, 3, 4, 4, 4, 5, 5, 6, 6, 6, 6, 6,
                                  7, 7, 7),
                     Npat     = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1),
                     Ntox     = c(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1),
                     Weight   = c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1),
                     PK       = c(2.82, 1.68, 2.87, 8.14, 4.39, 13.30, 12.20, 79.70, 47.20,
                                  41.00, 297.00, 164.00, 401.00, 766.00, 592.00, 769.00,
                                  577.00, 834.00, 1160.00, 1160.00)/median(c(2.82, 1.68, 2.87, 
                                  8.14, 4.39, 13.30, 12.20, 79.70, 47.20, 41.00, 297.00, 164.00, 
                                  401.00, 766.00, 592.00, 769.00, 577.00, 834.00, 1160.00, 
                                  1160.00)),
                     X        = NULL
                     )
                     
Doses <- c(0.1, 0.3, 1.0, 3.0, 10.0, 30.0, 50.0)


# Analysis: Fixed effect model assuming monotone relationship between PK and Doses

fit.BLRM.PK.Fixed.mono <- SingleBLRM.PK(Data           = Data.Dose.PK,
                                        Agent          = "DRUG 1",
                                        Doses          = Doses,
                                        Doses.Label    = as.character(Doses),
                                        DoseRef        = 270,
                                        Prior.logAB    = list(c(log(0.50), 0, 2, 1, 0)),
                                        pmix.logAB     = c(1),
                                        monotone.logAB = TRUE,
                                        Prior.logAB.PK = list(c(1, 0, 3, 1, 0)),
                                        pmix.logAB.PK  = c(1),
                                        monotone.PK    = TRUE,
                                        Prior.sd.PK    = c(0.25, log(2)/1.96),
                                        random.model   = FALSE,
                                        Prior.beta     = rbind(NULL,c(0, 10)),
                                        Prior.Only     = FALSE,
                                        Pcutoffs       = c(0.16, 0.33),
                                        New.cohortsize = 3,
                                        Outfile        = "out_BLRMPK_Fixed_mono",
                                        MCMC           = c(5000, 10000, 4, 1),
                                        DIC            = TRUE,
                                        Rnd.seed       = 1452671,
                                        RSeed.WinBUGS  = 1367
                                        )

}

}
\references{

}
\seealso{
\code{\link{SingleBLRM}}
}
