% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleBLRM.R
\name{SingleBLRM}
\alias{SingleBLRM}
\title{SingleBLRM Function}
\usage{
SingleBLRM(
  Data = NULL,
  Agent = NULL,
  Doses = NULL,
  DoseRef = NULL,
  Prior.logAB = NULL,
  pmix.logAB = c(1),
  Prior.Only = FALSE,
  Pcutoffs = c(0.16, 0.33),
  New.cohortsize = 3,
  Outfile = "Out",
  Plot = TRUE,
  Table = TRUE,
  int.crit = c(1, 1, 0.25),
  pars = c("logAlphaBeta", "pTox", "pCat", "pred.Tox"),
  MCMC = c(2500, 7500, 4, 1),
  Warn = TRUE,
  DIC = TRUE,
  progress.bar = "text",
  Rnd.seed = 1452671,
  RSeed.WinBUGS = 1367
)
}
\arguments{
\item{Data}{dose-DLT data: a list object contains DoseAdm, Npat, Ntox, and Weight. DoseAdm, Npat, 
 and Ntox are vectors of doses administered, Npat number of patients treated and the number of
dose-limiting toxicitites observed per cohort. The vector Weight represents the relevance of the cohort
for current trial. If Weight remains unspecified, then it will equal 1 for each cohort.}

\item{Agent}{name of the Agent under investigation.}

\item{Doses}{a vector of Doses for which inferential results will be provided.}

\item{DoseRef}{refernce dose for BLRM.}

\item{Prior.logAB}{prior for the model parameters Bayesian Logistic Regession Model (a list object).
each element of the list is vector with five elements. 
(\eqn{m_\alpha}, \eqn{m_\beta}, \eqn{s_\alpha}, \eqn{s_\beta}, \eqn{Corr(\alpha,\beta)}). These five 
parameters represent prior mean, sd and correlation of the bivariate normal prior for 
(\eqn{\log(\alpha)}, \eqn{\log(\beta)}). If no mixture prior is used, the list contains a single vector. 
Otherwise, the list  contains a number of five component vectors. The number of vectors are equal to number
of mixture components.}

\item{pmix.logAB}{a vector with apriori mixture probabilities. The number of components of pmix.logAB.
must be same as number of components in Prior.logAB.See the package vignettes for details.}

\item{Prior.Only}{logical: TRUE means data is ignored; if FALSE (the default), then data is taken
into account.}

\item{Pcutoffs}{cutoffs to define the intervals for the true DLT probability. The default is
(0.16, 0.33) defining three categories: underdosing ({[}0--0.16{]}), target ((0.16--0.33)), and 
overdosing ({[}0.33--1.00{]}).}

\item{New.cohortsize}{number of subjects in a new cohort, for which the predictive distribution 
of the number of DLT's will be computed.}

\item{Outfile}{name of the output file.}

\item{Plot}{logical: if TRUE (default) plots will be generated.}

\item{Table}{logical: if TRUE (default) the output tables will be generated.}

\item{int.crit}{vector of critical values for each interval defined by Pcutoffs (only used for
graphical output).}

\item{pars}{MCMC parameters to save for output summary.}

\item{MCMC}{four component vector of MCMC parameters n.burnin, n.iterations, n.chains, n.thin.
See \link{R2WinBUGS} for details.}

\item{Warn}{logical: if TRUE, the convergence warning will be printed.}

\item{DIC}{logical: if TRUE, compute deviance, pD, and DIC. See \link{R2WinBUGS} for details.}

\item{progress.bar}{type of progress bar. Possible values are “text”, “gui”, and “none”. Type “text”
is displayed on the R console. See \link{rjags} for details.}

\item{Rnd.seed}{Random seed for the R random stream. This is important for reproducibility.}

\item{RSeed.WinBUGS}{Random seed for the WinBUGS random stream. This is important for reproducibility.}
}
\value{
The function returns as a list the following main items:
\item{data}{The dataset analyzed}
\item{Priors}{summary of prior specifications}
\item{logAlphaBeta}{Posterior summary of (\eqn{\log(\alpha)}, \eqn{\log(\beta)})}
\item{logAlphaBeta.cor}{Posterior summary of the correlation between (\eqn{\log(\alpha)}, \eqn{\log(\beta)})}
\item{logAlphaEBeta}{Posterior summary of (\eqn{\log(\alpha)}, \eqn{\beta})}
\item{P}{Posterior summary of DLT probabilities}
\item{Pcat}{Posterior summary of DLT interval probabilities}
\item{PcatP}{P and Pcat combined}
\item{pred.Tox}{Predicted probabilities for the number of DLT's for a cohort of size New.cohortSize}
\item{Post.pmix}{Posterior mixture weigths (for mixture prior)}
\item{Rhat}{Gelman-Rubin MCMC convergence diagnostics}
\item{MCMC}{MCMC specifications}
\item{R2WB}{WinBUGS object with summaries of MCMC analysis (see \link{R2WinBUGS})}

In addition the following files will be created:
\item{Outfile}{a text output file with summary outputs}
\item{Outfile_P}{estimates and 95\% interval plots for DLT rates (in pdf format)}
\item{Outfile_Pcat}{interval plots for DLT rates (in pdf format)}
}
\description{
This function performs analysis of dose limiting toxicity (DLT) data with Bayesian logistic regression (BLRM) for 
single agent Phase I Oncology trial
}
\details{
Statistical model:
\deqn{r_j|n_j \sim \mbox{Binomial} (\pi(d_j), n_j)} 
\deqn{logit(\pi(d_j)) = \log (\alpha) + \beta \log (d_j/d^*).}

Here, \eqn{r_j} and \eqn{n_j} are number of patients treated and number of DLT's observed 
at dose level \eqn{d_j}. Moreover, \eqn{\alpha} is the odds, \eqn{\pi/(1 - \pi} of a DLT at \eqn{d^*}, an 
arbitrary scaling dose. \eqn{\beta > 0} is the increase in the log-odds of a DLT by a unit increase in 
log-dose. For instance, doubling the dose from \eqn{d} to \eqn{2d} will increase the odds of a DLT by 
the factor \eqn{2^\beta}, from \eqn{\pi_d/(1 - \pi_d)} to \eqn{2^\beta \pi_d/( 1- \pi_d)}. 
Neuenschwander et al. 2008 have suggested bivariate normal prior for \eqn{(\log(\alpha), \log(\beta))}.  

\deqn{(\log(\alpha), \log(\beta)) \sim N_2 (m_\alpha, m_\beta, s_\alpha, s_\beta, Corr(\alpha,\beta))} 

Furthermore, mixture priors can be used for \eqn{(\log(\alpha), \log(\beta))},
\deqn{(\log(\alpha), \log(\beta)) \sim \sum \omega_i N_2 ({m_\alpha}_i, {m_\beta}_i, {s_\alpha}_i, {s_\beta}_i, Corr(\alpha_i,\beta_i)); \;\; \sum \omega_{i} = 1}

Details on prior derivation can be found in the book chapter by Neuenschwander et al 2015.
}
\examples{
\dontrun{
## Example 1: Analysis of phase I single agent data: prior only 

fit.BLRM.single.P <- SingleBLRM(Data           = NULL,
                                Agent          = "DRUG 1",
                                Doses          = c(1, 2, 4, 8, 15, 30, 40), 
                                DoseRef        = 10,
                                Prior.logAB    = list(c(log(0.5), 0, 2, 1, 0)),
                                pmix.logAB     = c(1),
                                Prior.Only     = TRUE,
                                Pcutoffs       = c(0.16, 0.33),
                                New.cohortsize = 3,
                                Outfile        = "Output_single_prior",
                                MCMC           = c(5000, 10000, 4, 1),
                                DIC            = FALSE,
                                Rnd.seed       = 1452671,
                                RSeed.WinBUGS  = 1367
                                )
   
   
   
## Example 2: Analysis of phase I single agent data: analysis of 6 cohort data 

fit.BLRM.single <- SingleBLRM(Data = list(DosesAdm = c(1, 2, 4, 8, 15, 30),
                                          Npat     = c(5, 6, 5, 9,  8,  4),
                                          Ntox     = c(0, 0, 0, 0,  1,  3),
                                          Weight   = c(1, 1, 1, 1,  1,  1)
                                          ),
                              Agent          = "DRUG 1",
                              Doses          = c(1, 2, 4, 8, 15, 30, 40), 
                              DoseRef        = 10,
                              Prior.logAB    = list(c(log(0.5), 0, 2, 1, 0)),
                              pmix.logAB     = c(1),
                              Prior.Only     = FALSE,
                              Pcutoffs       = c(0.16, 0.33),
                              New.cohortsize = 3,
                              Outfile        = "Output_single",
                              MCMC           = c(5000, 10000, 4, 1),
                              DIC            = FALSE,
                              Rnd.seed       = 1452671,
                              RSeed.WinBUGS  = 1367
                              )
         
                              
                                                                        

## Example 3: Analysis of phase I single agent data: analysis of 6 cohort data 
##            together with down-weighted data from other trial#' 

fit.BLRM.single.D <- SingleBLRM(Data = list(DosesAdm = c(1, 2, 4, 8, 15, 30,   8,  15),
                                            Npat     = c(5, 6, 5, 9,  8,  4,   6,   6),
                                            Ntox     = c(0, 0, 0, 0,  1,  3,   0,   2),
                                            Weight   = c(1, 1, 1, 1,  1,  1, 0.5, 0.5)
                                            ),
                                Agent          = "DRUG 1",
                                Doses          = c(1, 2, 4, 8, 15, 30, 40), 
                                DoseRef        = 10,
                                Prior.logAB    = list(c(log(0.5), 0, 2, 1, 0)),
                                pmix.logAB     = c(1),
                                Prior.Only     = FALSE,
                                Pcutoffs       = c(0.16, 0.33),
                                New.cohortsize = 3,
                                Outfile        = "Output_single_dwght",
                                MCMC           = c(5000, 10000, 4, 1),
                                DIC            = FALSE,
                                Rnd.seed       = 1452671,
                                RSeed.WinBUGS  = 1367
                                )
                                
       
                                
                                                         
## Example 4: Analysis of phase I single agent data with a mixture prior.
##            A 0.5-0.5 mixture of prior distribution is applied

fit.mixture <- SingleBLRM(Data = list(DosesAdm = c(1, 2, 4, 8, 15, 30),
                                      Npat     = c(5, 6, 5, 9,  8,  4),
                                      Ntox     = c(0, 0, 0, 0,  1,  3),
                                      Weight   = c(1, 1, 1, 1,  1,  1)
                                      ),
                          Agent          = "DRUG 1",
                          Doses          = c(1, 2, 4, 8, 15, 30, 40), 
                          DoseRef        = 10,
                          Prior.logAB    = list(c(log(0.5), 0, 2, 1, 0),
                                                c(0, 0, 1, 1, 0.5)),
                          pmix.logAB     = c(0.5, 0.5),
                          Prior.Only     = FALSE,
                          Pcutoffs       = c(0.16, 0.33),
                          New.cohortsize = 3,
                          Outfile        = "Output_mixture",
                          MCMC           = c(5000, 10000, 4, 1),
                          DIC            = FALSE,
                          Rnd.seed       = 1452671,
                          RSeed.WinBUGS  = 1367
                          )                               
}


}
\references{
Neuenschwander B, Branson M, Gsponer T. Critical aspects of the Bayesian approach to phase I
cancer trials. Stat Med. 2008;27(13):2420-39.

Neuenschwander B, Matano A, Tang Z, Roychoudhury S,Wandel S, Bailey S. A Bayesian Industry
Approach to Phase I Combination Trials in Oncology. Statistical Methods for Drug Combination
Studies. Taylor & Francis, 2015.
}
\seealso{
\code{\link{SingleBLRMSim}}, \code{\link{MAP.BLRM}}.
}
