% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleBLRMPKSimDecisionRule.R
\name{SingleBLRMPKSimDecisionRule}
\alias{SingleBLRMPKSimDecisionRule}
\title{SingleBLRMSimDecisionRule Function}
\usage{
SingleBLRMPKSimDecisionRule(
  MaxN,
  mnMTD,
  mnTOT,
  mnTAR,
  EWOC = 0.25,
  Escalation = c("dose_max", "maximal", "target"),
  doseMaxEsc = 2
)
}
\arguments{
\item{MaxN}{Maximum number of patients in each trial. The default is 60. See Details}

\item{mnMTD}{At least these many patients have been treated at MTD. The default is 6. See Details}

\item{mnTOT}{A minimum number of these many patients should have already treated in the trial. 
The default is 15. See Details}

\item{mnTAR}{Probability thresold of target toxicity. The default is 0.5. See details}

\item{EWOC}{Escalation with overdose control criterion. The default is 0.25. See details}

\item{Escalation}{Escalation The escalation method to be used. Possible methods -- "maximal":
recomended, maxiumum dose which satisfies EWOC criteria; "target": dose which has 
highest target toxicity probability; "dose_max": maximum dose, doses will not be skipped}

\item{doseMaxEsc}{Either a scaler or a vector of length of number of provisional doses. If a scaler
\eqn{k} then it used assumed that for all doses escaltion upto \eqn{k}-fold is allowed. If a vector
then for each dose the corresponding fold is allowed to escalate. The first element is set to 2.}
}
\description{
This function performs escalation and decision rule trial simulation with 
single agent
}
\details{
At any trial dose escalation is achieved satisfying the escalation with overdose control 
(EWOC) principle. This is implemented by setting the target toxicity interval to Pcutoffs, with
an overdose control criterion \eqn{P(\pi_d > Pcutoffs[2]) < EWOC}. 

Dose escalations continue until declaration of the maximum tolerated dose (MTD) \eqn{d}. 
This dose must meet the following conditions:

1. At least mnMTD patients have been treated at dose \eqn{d}.

2. This dose satisfies one of the following conditions:

i. The probability of targeted toxicity at \eqn{d} exceeds mnTAR.

ii. A minimum number of mnTOT pateints should have already been treated in the trial.
}
\seealso{
\code{\link{SingleBLRMPKSim}}
}
