% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleBLRMSim.R
\name{SingleBLRMSim}
\alias{SingleBLRMSim}
\title{SingleBLRMSim Function}
\usage{
SingleBLRMSim(
  Design = NULL,
  nTrials = 1000,
  TrueProbs = NULL,
  StartDose = NULL,
  CohortRnd = list(c(3, 4, 5, 6), c(0.8, 0.1, 0.05, 0.05)),
  Histdatf = 0,
  Histdata = NULL,
  Pcutoffs = c(0.16, 0.33),
  Nsim = c(2500, 5000, 2, 1),
  Decision = SingleBLRMSimDecisionRule(MaxN = 60, mnMTD = 6, mnTOT = 15, mnTAR = 0.5,
    EWOC = 0.25, Escalation = "maximal", doseMaxEsc = 2),
  Outfile = "Sims",
  Out.summary = "sims_summary.txt",
  RndSeed = 12
)
}
\arguments{
\item{Design}{Design specification: contains required parameters for \code{\link{SingleBLRM}}. This incudes provisional
dose levels, reference dose, and prior specifications. See \code{\link{SingleBLRM}} for more details.}

\item{nTrials}{Number of simulated trials.}

\item{TrueProbs}{a vector of true toxicity  probabilities for different dose levels.}

\item{CohortRnd}{Cohort size (Random). A list with the first element a vector of the available
cohort size(s), and the second element a vector of the same length with the 
respective selection probability of the cohort size. For example, when set to  
list(c(3, 2), c(0.7, 0.3)), then the cohortsize is chosen as 3 or 2 at random 
with probability 0.7 or 0.3 respectively. If a fixed cohort size is desired use 
e.g. CohortRnd = list(c(3), c(1))}

\item{Histdatf}{logical: if TRUE, historical data is used (optional)}

\item{Histdata}{a list with named elements  DosesAdm1, DosesAdm2, Npat, Ntox, Weight containing the
historical data (optional)}

\item{Pcutoffs}{cutoffs to define the intervals for the true DLT probability. The default is
(0.16, 0.33) defining three categories: underdosing ({[}0--0.16{]}), target ((0.16--0.33)), and 
overdosing ({[}0.33--1.00{]}).}

\item{Nsim}{four component vector of MCMC parameters n.burnin, n.iterations, n.chains, n.thin.
See \link{R2WinBUGS} for details.}

\item{Decision}{Function which decides after each simulated cohort how to proceed}

\item{Outfile}{A R object saves all simulations output objects}

\item{Out.summary}{A txt file with simulation summary}

\item{RndSeed}{Seed used for the random number generator. Required parameter}
}
\value{
Input arguments are 
\item{Design}{The input data and prior information}
\item{nTrials}{Number of simulated trials}
\item{TrueProbs}{True toxicity vector}
\item{StartDose}{Start dose for the escalation trial}
\item{Pcutoffs}{Target toxicity interval}
\item{CohortRnd}{A list of size two with first quantity is the vector of cohortsize and second quantity is 
                 the vector of probabilities assigned to each cohortsize. For example, when set to 
                 list(c(3, 2), c(0.7, 0.3)), then the cohortsize is chosen as 3 or 2 at random 
                 with probability 0.7 or 0.3 respectively.}
\item{Decision}{}
\item{Outfile}{Name of the output file}
\item{Out.summary}{Name of the summary output file}
\item{RndSeed}{The seed which was fixed to reproduce the result}

Output arguments are 
\item{DLTMATRX}{the occurrences of DLT's per dose per trial}
\item{MTDSIM}{a binary matrix which is only TRUE when MTD has been observed in the trial}
\item{NNMATRX}{the cohort sizes per dose per trial}
\item{res.sim$simul}{descrioption of each trial -- Trial no. (Trial), Cohort no. (Cohort), 
Number of toxicity (Ntox), Number of patients (Npat), Weight, Current dose (DoseAdm), 
underdose probability (under), probability of target toxicity (target), 
overdose probability (over), probability of the current dose (true.prob), 
suggestion for the next dose, if any, (nextDoseAdm), underdose probability of the next 
dose (nextunder), target toxicity probability of the next dose (nexttarget), 
probability of the next dose (nextprob), a binary variable which is only 1 (TRUE) 
if the current dose is declared as the MTD, if the first dose is so toxic that 
\eqn{\Pr(\pi_d > 0.33) \ge EWOC} then the trial is stooped and too.toxic variable is set to 1, 
the Rhat.max information, crash = 1 if the code is crashed, the seed number for R (Rseed) 
and JAGS (WBseed)}
\item{res.sim$simsummary}{description of those cohorts within those trials 
in which the corresponding dose is the MTD}
\item{sim.sum}{simulation summary -- Proportion of patients in target dose, 
Proportion of patients in over dose, Proportion of patients in under dose,
Proportion of trials with MTD within target dose region, 
Proportion of trails with MTD within over dose region, 
Proportion of trails with MTD within under dose region, 
Proportion of trials stopped due to too toxicity at the start dose, 
Average sample size}

In addition, an R object output and a text output are generated containing all of the above.
}
\description{
This function performs Phase I Oncology trial simulation for single agent using Bayesian 
Logistic Regression Model (BLRM) and Escalation with Overdose Criteria (EWOC)
}
\details{
See vignettes. browseVignettes("OncoPh1BLRM")
}
\examples{
\dontrun{
# Design Parameters
Design.ex <- list(Agent1      = "Agent 1",
                  Doses       = c(0.5, 0.75, 1.0, 1.5, 2, 2.5),
                  DoseRef     = 2.5,
                  Prior.logAB = list(c(log(0.5), 0, 2, 1, 0)),
                  pmix.logAB  = c(1)
)

# Simulation Scenario 1: using the mean from prior
TrueProbs.prior <- c(0.1799316, 0.2115959, 0.2177317, 0.2480946, 0.2825048, 0.3096862)

#Running Simulation
sim.args <- list(Design      = Design.ex,
                 nTrials     = 10,
                 TrueProbs   = TrueProbs.prior,
                 StartDose   = 0.5,
                 Pcutoffs    = c(0.16, 0.33),
                 CohortRnd   = list(c(3), c(1)),
                 Decision    = SingleBLRMSimDecisionRule(MaxN = 60, mnMTD = 6, 
                                                         mnTOT = 15, mnTAR = 0.5, 
                                                         EWOC = 0.25, 
                                                         Escalation  = "maximal", 
                                                         # maxiumum dose which satisfies EWOC criteria
                                                         doseMaxEsc  = c(2, 4, 3, 2, 2, 1.5)
                                                         ),
                 Outfile     = "Output_Simulation",
                 Out.summary = "Output_Summary_Simulation.txt",
                 RndSeed     = 123
)

# Setting up back-end for parallel computing (DO NOT CHANGE)
library(doParallel)
require(doRNG)
nCores <- detectCores()  # number of cores
cl <- makeCluster(nCores) 
registerDoParallel(cl)
registerDoRNG(seed = 123, once = FALSE) #SEED1 : To initiate enviorment. 
# THIS IS THE IMPORTANT STEP FOR PARALLELISM AND REPRODUCIBILITY
getDoParWorkers()

simulation.single <- do.call(SingleBLRMSim, sim.args)

# Stop all clusters (DO NOT CHANGE)
OncoPh1BLRM:::unregister()
stopImplicitCluster() 
}


}
\references{
#'Neuenschwander B, Branson M, Gsponer T. Critical aspects of the Bayesian approach to phase I
cancer trials. Stat Med. 2008;27(13):2420-39.

Neuenschwander B, Matano A, Tang Z, Roychoudhury S,Wandel S, Bailey S. A Bayesian Industry
Approach to Phase I Combination Trials in Oncology. Statistical Methods for Drug Combination
Studies. Taylor & Francis, 2015.
}
\seealso{
\code{\link{SingleBLRM}}, \code{\link{SingleBLRMSimDecisionRule}}
}
